# backend/docs/swagger.yaml
basePath: /
definitions:
...
  app.AdminUserCreateData:
    properties:
      email:
        type: string
      faculty_id:
        type: integer
      full_name:
        type: string
      group_id:
        type: integer
      job_title:
        type: string
      lectern_id:
        type: integer
      login:
        type: string
      password:
        type: string
      role:
        type: string
    type: object
  app.AdminUserUpdateData:
    properties:
      email:
        type: string
      faculty_id:
        type: integer
      lectern_id:
        type: integer
      login:
        type: string
      password:
        type: string
      role:
        type: string
      status:
        type: string
      student_id:
        type: integer
      teacher_id:
        type: integer
      user_id:
        type: integer
    type: object
...
  app.ConfirmEmailData:
    properties:
      code:
        type: string
    type: object
...
  app.RequestEmailData:
    properties:
      email:
        type: string
    type: object
...
  app.Response:
    properties:
      error:
        type: string
      id:
        type: string
      ok:
        type: boolean
      result: {}
    type: object
...
  app.TwoFaCodeData:
    properties:
      code:
        type: string
    type: object
...
  app.UserAgreementDecisionData:
    properties:
      decision:
        type: string
      version:
        type: string
    type: object
...
info:
  contact: {}
  description: Backend API for e-journal, attendance links and invite registration.
  title: EJournal Backend API
  version: "1.0"
paths:
...
  /api/admin/users:
    get:
      description: Returns users with pagination, search and filters. Admin only.
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Search by login or email
        in: query
        name: search
        type: string
      - description: Role
        in: query
        name: role
        type: string
      - description: Status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: List users for admin
      tags:
      - admin users
    post:
      consumes:
      - application/json
      description: Creates a user and the required role profile in one transaction.
        Admin only.
      parameters:
      - description: User payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app.AdminUserCreateData'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: Create user
      tags:
      - admin users
  /api/admin/users/{user_id}:
    delete:
      description: Soft deletes a user by setting status=archived. Admin only.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: Archive user
      tags:
      - admin users
    get:
      description: Returns one user. Admin only.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: Get user for admin
      tags:
      - admin users
    patch:
      consumes:
      - application/json
      description: Updates user fields, status or role. Admin only.
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: User payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app.AdminUserUpdateData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/app.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - admin users
...
  /api/user/2fa/disable:
    post:
      description: Disables 2FA and removes the stored TOTP secret for the current
        user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: Disable 2FA
      tags:
      - profile
  /api/user/2fa/generate:
    post:
      description: Generates a TOTP secret and QR code for the current user after
        validating email confirmation code.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: Generate 2FA setup
      tags:
      - profile
  /api/user/2fa/request-enable:
    post:
      description: Sends a confirmation code to user's bound email before allowing
        2FA QR code generation.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: Request 2FA enablement code
      tags:
      - profile
  /api/user/2fa/verify:
    post:
      consumes:
      - application/json
      description: Verifies a TOTP code and enables 2FA for the current user.
      parameters:
      - description: TOTP code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app.TwoFaCodeData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: Verify and enable 2FA
      tags:
      - profile
  /api/user/agreements/current:
    get:
      description: Returns the current agreement version and the latest user decision.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: Get current user agreement status
      tags:
      - profile
  /api/user/agreements/decision:
    post:
      consumes:
      - application/json
      description: Stores either acceptance or rejection of the current user agreement
        version.
      parameters:
      - description: Agreement decision
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app.UserAgreementDecisionData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: Record user agreement decision
      tags:
      - profile
...
  /api/user/email/bind/confirm:
    post:
      consumes:
      - application/json
      description: Confirms the email address using the code sent to the user.
      parameters:
      - description: Confirmation code
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app.ConfirmEmailData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: Confirm email binding
      tags:
      - profile
  /api/user/email/bind/request:
    post:
      consumes:
      - application/json
      description: Sends a confirmation code to the requested email address.
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/app.RequestEmailData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: Request email binding
      tags:
      - profile
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"